+++ src/inspect/sxsignature.hpp
