+++ src/policy/policy.hpp
@@ -47,7 +47,7 @@
 #include <baseproxy.hpp>
 #include <ext/libcidr/cidr.hpp>
 #include <ranges.hpp>
-#include <addrobj.hpp>
+#include <policy/addrobj.hpp>
 
 #include <sobject.hpp>
