+++ src/shm/shmauth.cpp
@@ -38,7 +38,7 @@
 */
 
 
-#include <authfactory.hpp>
+#include <policy/authfactory.hpp>
 #include <log/logger.hpp>
 #include <cfgapi.hpp>
